+++ xdev/cli/docstr_stubgen.py
@@ -1,3 +1,4 @@
+# mypy: ignore-errors
 """
 Script for auto-generating pyi type extension files from google-style
 docstrings.
@@ -5,6 +6,9 @@ docstrings.
 This is a work in progress, but ultimately the goal is to be able to express
 concise typing information in docstrings and then explicitly expose that to
 Python.
+
+It seems that mypy updates break this code extremely frequently. It is hard to
+keep this maintained.
 
 Requirements:
     pip install mypy autoflake yapf
@@ -37,9 +41,11 @@ Ignore:
     print(lib.current_sourcecode())
 """
 
+
 try:
-    from mypy.stubgen import (StubGenerator, find_self_initializers, FUNC, EMPTY,
+    from mypy.stubgen import (ASTStubGenerator, find_self_initializers, FUNC, EMPTY,
                               METHODS_WITH_RETURN_VALUE,)
+    from mypy.stubgen import (is_none_expr)
     from mypy.nodes import (
         # Expression, IntExpr, UnaryExpr, StrExpr, BytesExpr, NameExpr, FloatExpr, MemberExpr,
         # TupleExpr, ListExpr, ComparisonExpr, CallExpr, IndexExpr, EllipsisExpr,
@@ -65,7 +71,7 @@ try:
         has_yield_expression
     )
 except Exception:
-    StubGenerator = object
+    ASTStubGenerator = object
     FuncDef = None
     METHODS_WITH_RETURN_VALUE = []
 
@@ -218,6 +224,7 @@ def generate_typed_stubs(modpath):
     options = stubgen.Options(
         pyversion=defaults.PYTHON3_VERSION,
         no_import=True,
+        inspect=False,
         doc_dir='',
         search_path=[],
         interpreter=sys.executable,
@@ -236,7 +243,7 @@ def generate_typed_stubs(modpath):
     # generate_stubs(options)
 
     mypy_opts = stubgen.mypy_options(options)
-    py_modules, c_modules = stubgen.collect_build_targets(options, mypy_opts)
+    py_modules, pyc_modules, c_modules = stubgen.collect_build_targets(options, mypy_opts)
 
     # Collect info from docs (if given):
     sigs = class_sigs = None  # type: Optional[Dict[str, str]]
@@ -268,16 +275,17 @@ def generate_typed_stubs(modpath):
         # print(f'target={target}')
         files.append(target)
         with stubgen.generate_guarded(mod.module, target, options.ignore_errors, options.verbose):
-            stubgen.generate_stub_from_ast(mod, target, options.parse_only,
-                                           # options.pyversion,
-                                           options.include_private,
-                                           options.export_less)
+            # stubgen.generate_stub_from_ast(mod, target, options.parse_only,
+            #                                # options.pyversion,
+            #                                options.include_private,
+            #                                options.export_less)
 
             gen = ExtendedStubGenerator(mod.runtime_all,
                                         # pyversion=options.pyversion,
                                         include_private=options.include_private,
                                         analyzed=not options.parse_only,
                                         export_less=options.export_less)
+            gen.module = mod.module
             assert mod.ast is not None, "This function must be used only with analyzed modules"
             mod.ast.accept(gen)
             # print('gen.import_tracker.required_names = {!r}'.format(gen.import_tracker.required_names))
@@ -343,8 +351,8 @@ def delete_unpaired_pyi_files(modpath):
     Cleanup pyi files corresponding to renamed or removed py files.
     """
     import os
-    from xdev.cli import dirstats
-    walker = dirstats.DirectoryWalker(modpath, block_dnames=['__pycache__'])
+    import xdev
+    walker = xdev.DirectoryWalker(modpath, exclude_dnames=['__pycache__'])
     walker._walk()
     dangling_pyi_fpaths = []
     for node in walker.graph.nodes():
@@ -459,8 +467,7 @@ def postprocess_hacks(text, mod):
         text,
         filename='<stdin>',
         style_config=style,
-        lines=None,
-        verify=False)
+        lines=None)
     # print(text)
     return text
 
@@ -759,7 +766,7 @@ def hacked_typing_info(type_name):
     return result
 
 
-class ExtendedStubGenerator(StubGenerator):
+class ExtendedStubGenerator(ASTStubGenerator):
 
     def _hack_for_info(self, info):
         type_name = info['type']
@@ -771,7 +778,10 @@ class ExtendedStubGenerator(StubGenerator):
 
             results = hacked_typing_info(type_name)
             for typing_arg in results['typing_imports']:
-                self.add_typing_import(typing_arg)
+                # FIXME: mypy removed this API
+                # self.add_typing_import(typing_arg)
+                ...
+
             for line in results['import_lines']:
                 self.add_import_line(line)
             for hack in results['hacks']:
@@ -956,7 +966,8 @@ class ExtendedStubGenerator(StubGenerator):
                         doctype_str = info['type'].replace(' ', '')
                         if all(n not in doctype_str for n in {'None', 'Optional'}):
                             info['type'] = info['type'] + ' | None'
-                            self.add_typing_import('Union')
+                            # FIXME: mypy removed this API
+                            # self.add_typing_import('Union')
         # ------------------------------------------
 
         args: List[str] = []
@@ -1024,6 +1035,8 @@ class ExtendedStubGenerator(StubGenerator):
                             annotation += '={!r}'.format(arg_.initializer.value)
                         elif isinstance(arg_.initializer, mypy.nodes.NameExpr):
                             annotation += '={}'.format(arg_.initializer.name)
+                        elif isinstance(arg_.initializer, mypy.nodes.UnaryExpr):
+                            annotation += '={}'.format(arg_.initializer.expr.value)
                         else:
                             # fallback, unhandled default
                             print(f'todo: Unhandled arg_.initializer={type(arg_.initializer)}')
@@ -1054,21 +1067,24 @@ class ExtendedStubGenerator(StubGenerator):
             try:
                 self.add_abc_import('Generator')
             except AttributeError:
-                self.add_typing_import('Generator')
+                # FIXME: mypy removed this API
+                # self.add_typing_import('Generator')
+                ...
             yield_name = 'None'
             send_name = 'None'
             return_name = 'None'
             for expr, in_assignment in all_yield_expressions(o):
-                if expr.expr is not None and not self.is_none_expr(expr.expr):
-                    self.add_typing_import('Any')
+                if expr.expr is not None and not is_none_expr(expr.expr):
+                    # self.add_typing_import('Any')  # fixme
                     yield_name = 'Any'
                 if in_assignment:
-                    self.add_typing_import('Any')
+                    # self.add_typing_import('Any')  # fixme
                     send_name = 'Any'
             if has_return_statement(o):
-                self.add_typing_import('Any')
+                # self.add_typing_import('Any')  # fixme
                 return_name = 'Any'
-            generator_name = self.typing_name('Generator')
+            # generator_name = self.typing_name('Generator')
+            generator_name = "collections.abc.Generator"
             if return_parsed_docstr_info is not None:
                 yield_name = return_parsed_docstr_info[1]
             retname = f'{generator_name}[{yield_name}, {send_name}, {return_name}]'
@@ -1195,11 +1211,3 @@ def modpath_coerce(modpath_coercable):
         raise TypeError('{}'.format(type(modpath_coercable)))
     modpath = ub.util_import.normalize_modpath(modpath)
     return modpath
-
-if __name__ == '__main__':
-    """
-    CommandLine:
-        python -m xdev.cli.gen_typed_stubs
-    """
-    from xdev.cli.docstr_stubgen import DocstrStubgenCLI
-    DocstrStubgenCLI.main()