+++ CHANGELOG.md
@@ -4,7 +4,18 @@ We are currently working on porting this changelog to the specifications in
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Version 1.5.1 - Unreleased
+
+## Version 1.5.2 - Unreleased 
+
+### Added
+* Ported `kwutil.datetime`, `kwutil.timedelta`, and `kwutil.ensure_rng`
+* Added features to pyversion CLI
+
+### Changed
+* Change default of dirstats `parse_content` to False.
+
+
+## Version 1.5.1 - Released 2024-02-10 
 
 ### Added